Move `_parallelize` and `get_n_jobs` to `core/`
Moves functions `_parallelize` and `get_n_jobs` to
`core/_parallelize.py` and makes them importable on modular level.
